Add `rfc5077-client` to the list of tools.
rfc5077-client proposes some more advanced tests against a server or a
pool of servers. It will try to reuse sessions with and without
tickets and will query several time each IP of a pool of servers. Use
this if you want to check support of SSL session resume of a server or
a pool of servers.